refactor: guards
put each guard into its own file so that imports become re-writable with swc or babel plugins
