+++ src/index.ts
@@ -27,10 +27,22 @@ export * from './forEach';
 export * from './forEachObj';
 export * from './fromPairs';
 export * from './groupBy';
-export * from './guards';
 export * from './identity';
 export * from './indexBy';
 export * from './intersection';
+export * from './isArray';
+export * from './isBoolean';
+export * from './isDate';
+export * from './isDefined';
+export * from './isError';
+export * from './isFunction';
+export * from './isNil';
+export * from './isNot';
+export * from './isNumber';
+export * from './isObject';
+export * from './isPromise';
+export * from './isString';
+export * from './isTruthy';
 export * from './keys';
 export * from './last';
 export * from './map';
